Improve payment by transfer in Firefox. Add support for "Forever" services. Payment was changed to be a separate page.

+++ includes/functions.php
@@ -379,20 +379,20 @@ function pay_by_admin($user)
     return $db->last_id();
 }
 
-function add_bought_service_info($uid, $user_name, $ip, $method, $payment_id, $service, $server, $amount, $auth_data, $email, $extra_data, $forever=false)
+function add_bought_service_info($uid, $user_name, $ip, $method, $payment_id, $service, $server, $amount, $auth_data, $email, $extra_data)
 {
     global $heart, $db, $lang;
 
     // Dodajemy informacje o kupionej usludze do bazy danych
     $db->query($db->prepare(
         "INSERT INTO `" . TABLE_PREFIX . "bought_services` " .
-        "SET `uid`='%d', `payment`='%s', `payment_id`='%s', `service`='%s', " .
-        "`server`='%d', `amount`='%s', `auth_data`='%s', `email`='%s', `extra_data`='%s'",
-        array($uid, $method, $payment_id, $service, $server, $forever ? "Na zawsze" : $amount, $auth_data, $email, json_encode($extra_data))
+        "SET `uid` = '%d', `payment` = '%s', `payment_id` = '%s', `service` = '%s', " .
+        "`server` = '%d', `amount` = '%s', `auth_data` = '%s', `email` = '%s', `extra_data` = '%s'",
+        array($uid, $method, $payment_id, $service, $server, $amount, $auth_data, $email, json_encode($extra_data))
     ));
     $bougt_service_id = $db->last_id();
 
-    $ret = "brak";
+    $ret = $lang['none'];
     if ($email) {
         $message = purchase_info(array(
             'purchase_id' => $bougt_service_id,
@@ -410,6 +410,7 @@ function add_bought_service_info($uid, $user_name, $ip, $method, $payment_id, $s
     }
 
     $temp_server = $heart->get_server($server);
+    $amount = $amount != -1 ? "{$amount} {$service['tag']}" : $lang['forever'];
     log_info(newsprintf($lang['bought_service_info'], $service, $auth_data, $amount, $temp_server['name'], $payment_id, $ret, $user_name, $uid, $ip));
     unset($temp_server);
 
@@ -480,11 +481,12 @@ function delete_players_old_services()
     }
 
     // Usuwamy usugi ktre zwróciły true
-    $db->query($db->prepare(
-        "DELETE FROM `" . TABLE_PREFIX . "players_services` " .
-        "WHERE `id` IN (%s)",
-        array(implode(", ", $delete_ids))
-    ));
+    if( !empty($delete_ids))
+        $db->query($db->prepare(
+            "DELETE FROM `" . TABLE_PREFIX . "players_services` " .
+            "WHERE `id` IN (%s)",
+            array(implode(", ", $delete_ids))
+        ));
 
     // Wywołujemy akcje po usunieciu
     foreach($players_services as $player_service) {
